fix(B-451): clamp stem fade weights in fp32 on CPU
Root cause: _StreamingStemWriter called Tensor.clamp(min=...) on CPU float16 fade tensors. PyTorch CPU does not implement clamp_min for Half.

Verification: targeted stem writer test passed; registry path governance test passed; default gate no longer stops at B-451 and now fails at B-452 corrupt video pipeline message assertion. App live verification not run; B-451 remains code-fix-pending-live-verification.

+++ tests/test_services/test_stem_separator_audio_decode.py
@@ -69,10 +69,10 @@ def test_streaming_stem_writer_crossfades_without_full_accumulator(tmp_path: Pat
         channels=1,
         sample_rate=44100,
     )
-    chunk_1 = torch.tensor([[[1.0, 2.0, 3.0, 4.0, 5.0]]])
-    fade_1 = torch.tensor([1.0, 1.0, 1.0, 1.0, 0.0])
-    chunk_2 = torch.tensor([[[10.0, 20.0, 30.0, 40.0, 50.0]]])
-    fade_2 = torch.tensor([0.0, 1.0, 1.0, 1.0, 1.0])
+    chunk_1 = torch.tensor([[[1.0, 2.0, 3.0, 4.0, 5.0]]], dtype=torch.float16)
+    fade_1 = torch.tensor([1.0, 1.0, 1.0, 1.0, 0.0], dtype=torch.float16)
+    chunk_2 = torch.tensor([[[10.0, 20.0, 30.0, 40.0, 50.0]]], dtype=torch.float16)
+    fade_2 = torch.tensor([0.0, 1.0, 1.0, 1.0, 1.0], dtype=torch.float16)
 
     writer.write_chunk(chunk_1 * fade_1.unsqueeze(0).unsqueeze(0), fade_1, ["vocals"], 2, False)
     writer.write_chunk(chunk_2 * fade_2.unsqueeze(0).unsqueeze(0), fade_2, ["vocals"], 2, True)
